% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{armature_types_france}
\alias{armature_types_france}
\title{Arrowheads from Neolithic France \insertCite{Crema2014-jj,Saintot1998-de}{evoarchdata}}
\format{
A data frame with 9 rows and 20 variables:
\describe{
\item{Period}{chronologically distinct phases}
\item{...}{the various armature types. See the PDF in data-raw/ for the paradigmatic classification scheme for the armatures data}
}
}
\source{
\url{https://doi.org/10.1016/j.jas.2014.07.014}
}
\usage{
armature_types_france
}
\description{
Counts of armature types from the Clairvaux and Chalain sites
(Pétrequin, 1986, 1989, 1997; Pétrequin and Bailly, 2004) in the Jura region
of southeast France. The dataset comprises a total of 280 arrowheads
subdivided into nine chronologically distinct phases (see details).
}
\details{
The phases, as defined by \insertCite{Saintot1998-de;textual}{evoarchdata},
are dated as follows:
\itemize{
\item I: 3700–3600 BC
\item II: 3450 BC
\item III: 3200 BC
\item IV: 3100 BC
\item V: 3050–3010 BC
\item VI: 3010–2930 BC
\item VII: 2850–2750 BC
\item VIII: 2750–2600 BC
\item IX: 2600–1650 BC
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
