% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ceramics_hittite_turkey}
\alias{ceramics_hittite_turkey}
\title{Ceramic bowls from Boğazköy-Hattusa \insertCite{Steele2010-bq,Brantingham2010-yd}{evoarchdata}}
\format{
A data frame with 18 rows and 7 variables:
\describe{
\item{Phase}{chronological phases}
\item{Main bowl group}{Main bowl type groups}
\item{Ware A, etc.}{Ceramic fabric types}
...
}
}
\source{
\url{https://doi.org/10.1016/j.jas.2009.12.039}
}
\usage{
ceramics_hittite_turkey
}
\description{
Frequencies of rim sherds of the main bowl groups in each of the four wares,
by phase, from  Boğazköy-Hattusa, a Hittite site in Turkey.
}
\details{
Data from Table 1 of \insertCite{Steele2010-bq;textual}{evoarchdata} and
Table 2 of \insertCite{Brantingham2010-yd;textual}{evoarchdata}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
