% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ceramics_woodland_illinois}
\alias{ceramics_woodland_illinois}
\title{Illinois Woodland Ceramics \insertCite{Braun1977-hy,Braun1985-wv,Neiman1995-ar}{evoarchdata}}
\format{
A data frame with 106 rows and 15 variables:
\describe{
\item{Group}{Group}
\item{Motif}{Motif}
\item{Total}{Total}
\item{mm02_3, etc.}{Total}
...
}
}
\source{
\url{http://www.people.virginia.edu/~fn9r/anth588/WoodlandIllinoisData.html}
}
\usage{
ceramics_woodland_illinois
}
\description{
Data on ceramics from the Middle Woodland-Early Late Woodland periods in
the Central Midwest (Illinois). Includes stylistic element frequencies for
lip exterior decoration for five regions, by thickness class \insertCite{Braun1977-hy,Braun1985-wv,Neiman1995-ar}{evoarchdata}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
