% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{lithics_acheulean_virtual}
\alias{lithics_acheulean_virtual}
\title{Copying error in virtual Acheulean handaxes}
\format{
A data frame with 200 rows and 4 variables:
\describe{
\item{chain}{Number of the transmission chain.}
\item{place}{Place of the handaxe in the chain.}
\item{scale}{Size of the handaxe on an arbitrary scale.}
\item{condition}{Whether the handaxe is larger or smaller than the previous handaxe in the chain.}
}
}
\source{
\url{https://doi.org/10.1371/journal.pone.0048333.s002}
}
\usage{
lithics_acheulean_virtual
}
\description{
Experimental data on unintential copying error in cultural transmission. In
\insertCite{Kempe2012-ns;textual}{evoarchdata}'s experiment, participants
were split into 20 transmission chains and asked to copy the size of a
virtual Acheulean handaxe.
}
\references{
\itemize{
\item \insertRef{Kempe2012-ns}{evoarchdata}
}
}
\keyword{datasets}
