% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{lithics_lpalaeolithic_neurope}
\alias{lithics_lpalaeolithic_neurope}
\title{Late Palaeolithic Projectile Points from Northern Europe \insertCite{Riede2009-he}{evoarchdata}}
\format{
A data frame with 632 rows and 6 variables:
\describe{
\item{number}{Unique identifier of the specimen.}
\item{length}{Measured length of the specimen in mm.}
\item{length_total}{Estimated total length of the point in mm, if it is broken. If not broken, equal to \code{length}.}
\item{width}{Width of the specimen in mm.}
\item{thickness}{Thickness of the specimen in mm.}
\item{type}{Typological classification according to literature or the museum creator.}
}
}
\source{
\url{http://www.cecd.ucl.ac.uk/ceacb_files/datasets/Riede_SI.xls}
}
\usage{
lithics_lpalaeolithic_neurope
}
\description{
Data on the shape and size of lithic projectile points from the Late
Palaeolithic of Northern Europe.
}
\references{
\itemize{
\item \insertRef{Riede2009-he}{evoarchdata}
}
}
\keyword{datasets}
