% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{textiles_ikat_seasia}
\alias{textiles_ikat_seasia}
\title{Southeast Asian Warp Ikat Weaving \insertCite{Buckley2012-ye}{evoarchdata}}
\format{
A data frame with 36 rows and 111 variables:
\describe{
\item{tradition}{Name of the weaving tradition.}
\item{...}{\code{logical}. Presence/absence of coded characters. See details for the coding system.}
}
}
\source{
\url{https://figshare.com/articles/dataset/Investigating_Cultural_Evolution_Using_Phylogenetic_Analysis_The_Origins_and_Descent_of_the_Southeast_Asian_Tradition_of_Warp_Ikat_Weaving__/115810}
}
\usage{
textiles_ikat_seasia
}
\description{
Data on decorative motifs used in warp ikat weaving, a method of textile
manufactured widely used in Southeast Asia. Includes 36 weaving traditions,
which according to Buckley "correspond to broadly recognized cultural
groups with distinct languages or dialects".
}
\details{
See the supplementary information (\code{Motif_S1.pdf} and \code{TaxaChars_S1.pdf})
in \insertCite{Buckley2012-ye;textual}{evoarchdata} for an explanation of
the coding system used.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
