% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ceramics_hittite_turkey}
\alias{ceramics_hittite_turkey}
\title{Counts of ceramic bowl types from Boğazköy-Hattusa, a Hittite site in Turkey (Steele et al. 2010; Brantingham and Perreault 2010)}
\format{A data frame with 18 rows and 7 variables:
\describe{
  \item{Phase}{chronological phases}
  \item{Main bowl group}{Main bowl type groups}
  \item{Ware A, etc.}{Ceramic fabric types}
  ...
}}
\source{
\url{https://doi.org/10.1016/j.jas.2009.12.039}
}
\usage{
ceramics_hittite_turkey
}
\description{
Frequencies of rim sherds of the main bowl groups in each of the four wares, by phase, from  Boğazköy-Hattusa, a Hittite site in Turke
}
\details{
Data from Table 1 of

Steele, J., Glatz, C., Kandler, A., 2010. Ceramic diversity, random copying, and tests for selectivity in ceramic production. Journal of Archaeological Science 37, 1348–1358. https://doi.org/10.1016/j.jas.2009.12.039

Also Table 2 of

Brantingham, P. J., & Perreault, C. (2010). Detecting the effects of selection and stochastic forces in archaeological assemblages. Journal of Archaeological Science, 37(12), 3211-3225.
}
\keyword{datasets}
